+++ Services/ProjectsService.cs
@@ -1,4 +1,5 @@
-﻿using InternIntelligence_Portfolio.Dtos.Projects;
+﻿using System.Security.Claims;
+using InternIntelligence_Portfolio.Dtos.Projects;
 using InternIntelligence_Portfolio.Models;
 using InternIntelligence_Portfolio.Repos;
 
@@ -15,28 +16,46 @@ namespace InternIntelligence_Portfolio.Services
     public class ProjectsService : IProjectsService
     {
         private readonly IGenericRepo<Projects> _genericRepo;
-
-        public ProjectsService(IGenericRepo<Projects> genericRepo)
+        private readonly IHttpContextAccessor _httpContextAccessor;
+        public ProjectsService(IGenericRepo<Projects> genericRepo,
+                                    IHttpContextAccessor httpContextAccessor)
         {
             _genericRepo = genericRepo;
+            _httpContextAccessor = httpContextAccessor;
         }
+        private int GetCurrentUserId()
+        {
+            var claimsPrincipal = _httpContextAccessor.HttpContext.User;
+            var userIdClaim = claimsPrincipal.FindFirst(ClaimTypes.NameIdentifier);
 
+            if (userIdClaim != null && int.TryParse(userIdClaim.Value, out int userId))
+            {
+                return userId;
+            }
+
+            throw new UnauthorizedAccessException("User Must Be LoggedIn In Order To Use This Feature!");
+        }
         public List<Projects> GetAll()
         {
-            return _genericRepo.GetAll().Select(p => new Projects
-            {
-                Title = p.Title,
-                Description = p.Description,
-                GitHubUrl = p.GitHubUrl,
-                CreatedAt = DateTime.UtcNow
+            int userId = GetCurrentUserId();
+            return _genericRepo.GetAll()
+                                .Where(a => a.UserId == userId)
+                                 .Select(p => new Projects
+                                 {
+                                     Title = p.Title,
+                                     Description = p.Description,
+                                     GitHubUrl = p.GitHubUrl,
+                                     CreatedAt = DateTime.UtcNow
 
-            }).ToList();
+                                 }).ToList();
         }
 
         public Projects GetById(int id)
         {
+            int userId = GetCurrentUserId();
+
             var getProjects = _genericRepo.GetById(id);
-            if (getProjects == null)
+            if (getProjects == null || getProjects.UserId != userId)
             {
                 return null;
             }
@@ -45,25 +64,44 @@ namespace InternIntelligence_Portfolio.Services
         }
         public void Add(ProjectDtos projectDtos)
         {
+            int userId = GetCurrentUserId();
+
             var projects = new Projects()
             {
                 Title = projectDtos.Title,
                 Description = projectDtos.Description,
                 GitHubUrl = projectDtos.GitHubUrl,
-                CreatedAt = DateTime.UtcNow
+                CreatedAt = DateTime.UtcNow,
+                UserId = userId
             };
             _genericRepo.Add(projects);
         }
 
         public void Delete(int id)
         {
+            int userId = GetCurrentUserId();
+
             var getId = _genericRepo.GetById(id);
-            _genericRepo.Delete(getId);
+
+            if (getId != null && getId.UserId == userId)
+            {
+                _genericRepo.Delete(getId);
+            }
+            else
+            {
+                throw new UnauthorizedAccessException("Not authorized to delete this achievement");
+            }
         }
         public void Update(int Id, ProjectDtos projectDtos)
         {
+            int userId = GetCurrentUserId();
+
             var projects = _genericRepo.GetById(Id);
 
+            if (projects == null || projects.UserId != userId)
+            {
+                throw new UnauthorizedAccessException("Not authorized to update this achievement");
+            }
             projects.Title = projectDtos.Title;
             projects.Description = projectDtos.Description;
             projects.GitHubUrl = projectDtos.GitHubUrl;
